shell_funcs

+++ labs_god/lab4/shell_sort/shell_funcs.h
@@ -2,7 +2,7 @@
 #define SHELL_FUNCS_H
 
 // Прототипы. Реализация находится в shell_funcs.c
-void insertionSort(int array[], int size, int gap, int *compares, int *swaps);
-int *shellSort(int array[], int size, int *compares, int *swaps);
+void insertionSort(int array[], int size, int gap, long long *compares, long long *swaps);
+int *shellSort(int array[], int size, long long *compares, long long *swaps);
 
 #endif
